refactor: extract trace parsing into composable TraceExtractor class
- Extract trace parsing methods from Evaluator base class into new TraceExtractor
- Move _parse_turn_level, _parse_tool_level, and _parse_conversation_level to TraceExtractor
- TraceExtractor is initialized with EvaluationLevel and provides single extract() method
- Evaluator now uses TraceExtractor internally via composition pattern
- Update evaluator subclasses to call super().__init__() after setting evaluation_level
- Add comprehensive tests for TraceExtractor demonstrating composability
- Update existing tests to use TraceExtractor directly
- Add example demonstrating TraceExtractor composability patterns

This refactoring improves separation of concerns and enables:
- Independent use of trace extraction outside evaluators
- Composition of multiple extractors for different analysis purposes
- Easier testing and extension of trace parsing logic
- Better alignment with existing extractors module pattern
