+++ src/strands_evals/extractors/__init__.py
@@ -1,0 +1,3 @@
+from .trace_extractor import TraceExtractor
+
+__all__ = ["TraceExtractor"]
